Add `-stdlib=libc++` when `CXX` is not specified
This argument is not permissible with `gcc`. Also if a custom compiler
like `icc` or a non-system provided copy of `clang` is used, this flag
will also not work. In other words, only with the macOS provided `clang`
can we safely add the `-stdlib=libc++` flag. If users are changing the
compiler to one that for some reason can take this flag, they can just
as well set it themselves. Hence we only set when `CXX` is not
specified.