`METH_KEYWORDS` -> `METH_VARARGS | METH_KEYWORDS`
This appears to be an annoying bug that was introduced in the very
beginning of Python 3 and has been around ever since unfortunately. :(
The proposed workaround is to just use note a method can also handle
varargs. Though it looks like it is getting fixed as of this year. I
don't think we will have releases generally available for this fix in
the near term. So this patch should suffice. More details in the linked
references.

ref: https://bugs.python.org/issue11587 (original)
ref: https://bugs.python.org/issue15657 (duplicate, was fixed)