Handle unicode regularizer
On Python 3, the default string is of Unicode type, which caused this
comparison some issues. In particular, the length comparison was off as
the Unicode string may have more bytes than the equivalent ASCII string.
To fix that, just encode Unicode strings as ASCII byte strings in
Python. This way handling of them can proceed exactly as before. As it
is also possible to get Unicode values on Python 2, it doesn't hurt to
handle them appropriately as well.