Improve error checking in regularizer
In some cases a conversion may not go smoothly or may not be the
expected type. The return values in these cases will be `NULL`. If an
error is set twice though, this result in an interpreter crash. So try
to handle a few of these around string conversion to avoid crashing.
Also call `PyErr_Clear` before throwing exceptions where there may
already be a Python exception being unwound. If there is not one being
unwound, this is a no-op.