Convert `range` to `list` on Python 3
In Python 3, `range` behaves like `xrange`. So it does not eagerly
evaluate and return a `list`. Instead it remains an object that can be
iterated over and has a few other properties. This differs from Python 2
where `range` returns a `list`. To correct this incompatibility, simply
pass the result of `range` to a `list`. If the result is already a
`list` as it is on Python 2, this is a no-op. However if `range` returns
an iterable object like it is on Python 3, this will convert it to a
`list` giving the expected result seen on Python 2.