Get the `pvalue` from SciPy's `ks_2samp`
It appears that SciPy's `ks_2samp` has slightly different behavior on
Python 2 vs. Python 3. In particular on Python 2 it reduces to its
`pvalue` when performing comparisons to numeric values. However on
Python 3 this behavior does not appear to be supported. To address this
issue, simply get the `pvalue` attribute regardless. This yields the
correct behavior on both Python 2 and Python 3. More details in the
linked issue.

xref: https://github.com/scipy/scipy/issues/6099